Code cleanup, fix labels positioning
